Time: 210 ms (42.81%), Space: 155.9 MB (9.24%) - LeetHub

+++ 2762-continuous-subarrays/2762-continuous-subarrays.cpp
@@ -3,7 +3,7 @@ class Solution {
 public:
     long long continuousSubarrays(vector<int>& a) {
         ll n = a.size(), ans = 0;
-        multiset<int> s;
+        multiset<ll> s;
         for (int i=0,j=0; j<n; j++) {
             s.insert(a[j]);
             while (s.size()) {
